requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
Django==3.0.3
django-bootstrap4==1.1.1
django-crispy-forms==1.8.1
djangorestframework==3.11.0
psycopg2-binary==2.8.4
pytz==2019.3
soupsieve==1.9.5
sqlparse==0.3.0
